apps/api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
pytest==8.4.1
httpx==0.28.1
email-validator==2.2.0
